Add Cream (thanks to @loodaobao)

+++ protocols/alpha/src/index.ts
@@ -9,6 +9,8 @@ const config = dotenv.config({path: "../.env"}).parsed!
 const path = __filename.replace("src/index.ts", "").concat("script/nam");
 
 export default async function getAlphaDistribution(options: Options) {
+    if(!fs.existsSync("script")) return {}
+    
     fs.writeFileSync(path + "/configs/end_block.json", String(options.endBlock));
     fs.writeFileSync(path + "/data/sushi_blacklist.json", JSON.stringify(options.blacklistDistribution, null, 1))
